Make use of version catalogs and variants
