+++ buildSrc/settings.gradle.kts
@@ -1,0 +1,22 @@
+dependencyResolutionManagement {
+    repositories {
+        mavenCentral()
+        gradlePluginPortal()
+        mavenLocal {
+            include("org.openrndr")
+        }
+    }
+
+    versionCatalogs {
+        create("libs") {
+            from(files("../gradle/libs.versions.toml"))
+        }
+    }
+}
+
+pluginManagement {
+    repositories {
+        gradlePluginPortal()
+
+    }
+}
