+++ gradle.properties
@@ -1,4 +1,5 @@
 openrndr.tasks=true
+applicationMainClass=TemplateProgramKt
 kotlin.code.style=official
 #org.gradle.configuration-cache=true
 #org.gradle.configuration-cache.parallel=true
